added auth check for delete

+++ blog_home/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+
 from django.contrib.auth.models import User
 
 from .models import Posts
@@ -10,6 +12,8 @@ from .models import Posts
 def home(request):
     if request.method=="GET":
         posts = Posts.objects.all()
+        for post in posts:
+            print(post.id)
         context = {
             'Posts' : posts
         }
@@ -21,9 +25,14 @@ def home(request):
         post = Posts(author=author, title=title, content=content)
         post.save()
         return redirect('blog-home')
-    
+
+
 def delete_post(request,id):
     post = Posts.objects.filter(id=id).first()
-    post.delete()
-    return redirect('blog-home')
-    
+    if request.user == post.author:
+        post.delete()
+        return redirect('blog-home')
+    else: 
+        messages.info(request,f"You dont have the persmission to delete this post")
+        return redirect('blog-home')
+
